OS Linux: add complete swap monitoring using LLD.
Complete swap space monitoring using LLD, so system swap monitoring
is added only when the monitored system uses swap device(s).
During the implementation of the swap monitoring, I found some
difficulties about which you can read on:
https://support.zabbix.com/browse/ZBX-16821
This is an excellent example of how to add similar things which add
the set of items/triggers/graphs or not depends on the state of some
single resource.

Monitoring of the swap consists of:
- "SWAP:" new LLD
  It adds prototypes:
  - items: "in", "total", "out" and "used" with discard unchanged filter
  - "SWAP" and "SWAP::in/out" graphs
  - trigger: "SWAP::low space ({$ITEM.VALUE}% used)"
    the trigger is activated when more the "{$SWAP_HIGH}" percent of the swap
    is used with default "{$SWAP_HIGH}=80" value
- Screens:
  - "SWAP" added with "SWAP" and "SWAP::in/out" graphs

TODO: add similar modification for L2ARC monitoring and MySQL slave databases
connections.